restruct 0.5 based 1.6
用户get_order get_voucher接口单独写入user.py
修改相应url
